first test of lockfile subPackage
