+++ src/mod_Mods.go
@@ -11,6 +11,7 @@ import (
     "fmt"
     "io/ioutil"
     "path/filepath"
+    "lockfile"
 )
 
 type Mods struct {
@@ -24,6 +25,8 @@ type ModsResult struct {
 type ModsResultList struct {
     ModsResult []ModsResult `json:"mods"`
 }
+
+var fileLock lockfile.FileLock = lockfile.newLock()
 
 func newMods(destination string) (Mods, error) {
     var err error
